feat: Clean up client-side routing
