+++ 권현지/6주차/spa/src/pages/index.js
@@ -1,0 +1,17 @@
+import { goto } from "../router";
+
+export const renderInx= ()=>{
+  document.querySelector('#app').innerHTML = `
+  <h1>movie info</h1>
+  <form>
+    <input type="search" name="query" />
+    <button type="submit">search</button> 
+  </form>
+  `;
+  document.body.querySelector('form').addEventListener('submit', (event) => {
+    event.preventDefault();
+    console.log(event.target.query.value);
+    // location.href= `/search?query=${event.target.query.value}`
+    goto(`/search?query=${event.target.query.value}`, { push: true });
+  });
+}
