+++ 권현지/6주차/spa/src/pages/search.js
@@ -1,0 +1,6 @@
+export const renderSearch = ({searchParam})=>{
+  document.querySelector('#app').innerHTML = `
+  <h1>movie search</h1>
+  <p>${searchParam.query}</p>
+`
+}
