+++ 권현지/6주차/spa/src/routes.js
@@ -1,0 +1,7 @@
+import { renderInx } from "./pages/index";
+import { renderSearch } from "./pages/search";
+
+export const routes ={
+  '/':renderInx,
+  '/search' : renderSearch
+}
